+++ NEWS.md
@@ -1,3 +1,5 @@
+* Changed directory layout: All data are now in the new `input` subdirectory,
+  and all results are in the new `output` subdirectory (#93).
 * Changed directory layout: All scripts are now in the new `code` subdirectory (#85).
 
 v0.6-1 (2014-11-28)